Move mapp config `useStatistics` hook
This centralize the config to avoid duplication of code
