+++ src/apps/search-result/search-result.entry.tsx
@@ -8,7 +8,8 @@ import SearchResult from "./search-result";
 import GlobalUrlEntryPropsInterface from "../../core/utils/types/global-url-props";
 import { GlobalEntryTextProps } from "../../core/storybook/globalTextArgs";
 import { MappArgs } from "../../core/storybook/mappArgs";
-import usePageStatistics from "../../core/statistics/usePageStatistics";
+import { useStatistics } from "../../core/statistics/useStatistics";
+import { useEffectOnce } from "react-use";
 
 interface SearchResultEntryTextProps {
   addMoreFiltersText: string;
@@ -72,7 +73,10 @@ const SearchResultEntry: React.FC<SearchResultEntryProps> = ({
   pageSizeDesktop,
   pageSizeMobile
 }) => {
-  usePageStatistics();
+  const { sendPageStatistics } = useStatistics();
+  useEffectOnce(() => {
+    sendPageStatistics();
+  });
   // If a q string has been defined as a data attribute use that
   // otherwise use the one from the url query parameter.
   const { q: searchQuery } = getParams({ q });
